pretty sure this is working, but we need to tune different parameters
pretty sure this is modeling what we want it to model (e.g., nodes are forwarding each others' packets): we get reasonable throughput on a topology with 2 nodes, 100 m apart; 0 Mb/s in a topology with 2 nodes, 150 m apart; and reasonable throughput with 3 nodes, each 100 m apart (would not be possible if nodes weren't forwarding each others' packets, i think). The core task left, then, is to figure out how to tune the parameters at the physical layer to model the effective transmission range that we want.